Core: Start layout changes for removing kernel.cpp
