+++ src/map/main.cpp
@@ -1,7 +1,7 @@
 ﻿/*
 ===========================================================================
 
-  Copyright (c) 2023 LandSandBoat Dev Teams
+  Copyright (c) 2025 LandSandBoat Dev Teams
 
   This program is free software: you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
@@ -19,22 +19,12 @@
 ===========================================================================
 */
 
-#include "connect_server.h"
-
-// openssl applink.c prevents issues with debug vs release vs threaded/single threaded .dlls at runtime
-// apparently not an issue with linux
-#ifdef _WIN32
-#include <ms/applink.c>
-#endif
-
-// TODO: Standardize our running arguments for shutdown and thread signals
-std::atomic<bool> gRunFlag = true;
+#include "map_server.h"
 
 int main(int argc, char** argv)
 {
-    auto pConnectServer = std::make_unique<ConnectServer>(argc, argv);
+    auto pMapServer = std::make_unique<MapServer>(argc, argv);
 
-    // TODO: We don't need this
     while (pConnectServer->IsRunning())
     {
         pConnectServer->Tick();
