+++ src/map/map_server.cpp
@@ -1,7 +1,7 @@
 ﻿/*
 ===========================================================================
 
-  Copyright (c) 2010-2015 Darkstar Dev Teams
+  Copyright (c) 2025 LandSandBoat Dev Teams
 
   This program is free software: you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
@@ -19,31 +19,7 @@
 ===========================================================================
 */
 
-#ifndef _MAP_H
-#define _MAP_H
-
-#include "common/cbasetypes.h"
-
-#include "common/blowfish.h"
-#include "common/kernel.h"
-#include "common/md52.h"
-#include "common/mmo.h"
-#include "common/socket.h"
-#include "common/sql.h"
-#include "common/taskmgr.h"
-#include "common/xirand.h"
-
-#include <list>
-#include <map>
-
-#include "command_handler.h"
-#include "zone.h"
-
-/************************************************************************
- *                                                                       *
- *  Map's working session                                                *
- *                                                                       *
- ************************************************************************/
+#pragma once
 
 struct map_session_data_t
 {
@@ -92,42 +68,10 @@ struct map_session_data_t
     }
 };
 
-extern uint32 map_amntplayers;
-extern int32  map_fd;
-
-// 2.5 updates per second
-static constexpr float server_tick_rate = 2.5f;
-
-// Update every 400ms
-static constexpr float server_tick_interval = 1000.0f / server_tick_rate;
-
-// Check Trigger Areas every 200ms
-static constexpr float server_trigger_area_interval = server_tick_interval / 2.0f;
-
 typedef std::map<uint64, map_session_data_t*> map_session_list_t;
 extern map_session_list_t                     map_session_list;
-
-extern in_addr map_ip;
-extern uint16  map_port;
 
 extern inline map_session_data_t* mapsession_getbyipp(uint64 ipp);
 extern inline map_session_data_t* mapsession_createsession(uint32 ip, uint16 port);
 
-extern std::unique_ptr<SqlConnection> _sql;
-
-extern bool gLoadAllLua;
-
-//=======================================================================
-
-int32 recv_parse(int8* buff, size_t* buffsize, sockaddr_in* from, map_session_data_t*);                      // main function to parse recv packets
-int32 parse(int8* buff, size_t* buffsize, sockaddr_in* from, map_session_data_t*);                           // main function parsing the packets
-int32 send_parse(int8* buff, size_t* buffsize, sockaddr_in* from, map_session_data_t*, bool usePreviousKey); // main function is building big packet
-
-void map_helpscreen(int32 flag);
-
-int32 map_cleanup(time_point tick, CTaskMgr::CTask* PTask); // Clean up timed out players
 int32 map_close_session(time_point tick, map_session_data_t* map_session_data);
-
-int32 map_garbage_collect(time_point tick, CTaskMgr::CTask* PTask);
-
-#endif //_MAP_H
